requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
colorama==0.4.4
Flask==2.0.0
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.0
Jinja2==3.0.0
jsonpickle==2.0.0
MarkupSafe==2.0.0
requests==2.25.1
urllib3==1.26.4
Werkzeug==2.0.0
python-dotenv~=0.17.1